requirements.txt
Django==1.2
MySQL>=5.0
